requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0

# Code quality
flake8==6.0.0
black==23.7.0
isort==5.12.0

# Development tools
django-debug-toolbar==4.1.0
django-extensions==3.2.3

# Documentation
sphinx==7.1.0
sphinx-rtd-theme==1.3.0
